refactor: Game을 객체로 생성하여 불필요한 파라미터의 정보를 숨기고, 최소한의 로직으로 controller 구성
